pyproject.toml
[project]
name = "resume-agent-plugin"
version = "1.0.0"
description = "AI-powered Resume Analysis & Optimization Plugin for Claude Code"
requires-python = ">=3.12"
dependencies = [
    "pymupdf",
    "jsonschema",
    "python-docx",
    "typst",
    "fpdf2",
]

[dependency-groups]
dev = [
    "pytest",
]
